+++ src/pages/index.ts
@@ -1,5 +1,5 @@
-export * from './Stations';
-export * from './Sections';
-export * from './Lines';
-export * from './Login';
-export * from './SignUp';
+export * from './StationsPage';
+export * from './SectionsPage';
+export * from './LinesPage';
+export * from './LoginPage';
+export * from './SignUpPage';
